requirements.txt
# Core dependencies
fastapi>=0.110.0
uvicorn[standard]>=0.28.0
pydantic>=2.6.0
python-dotenv>=1.0.0

# GraphQL client
gql>=3.4.0
requests>=2.31.0

# Language model integration
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.12.0

# Logging and observability
structlog>=24.1.0
python-json-logger>=2.0.7

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
pytest-cov>=4.1.0

# Performance and utilities
ujson>=5.8.0
cachetools>=5.3.2
aiofiles>=23.2.1
tenacity>=8.2.3

# Security
python-jose[cryptography]>=3.3.0

